Version 2.1
Some bug fixes. Vimball has been checked and should work Ok now.

However, in windows "checktime" does not work and the "pedit!" workaround is problematic when used with tabs and therefore has been commented out. As it is Windows users need to manually reload - or uncomment the "pedit!" workaround and not use tabs.